start inserting data into table and refactor controller to support both designs
