+++ app/controllers/case_contacts_controller.rb
@@ -1,30 +1,15 @@
 # frozen_string_literal: true
 
 class CaseContactsController < ApplicationController
+  include LoadsCaseContacts
+  
   before_action :set_case_contact, only: %i[edit destroy]
   before_action :set_contact_types, only: %i[new edit create]
   before_action :require_organization!
   after_action :verify_authorized, except: %i[leave]
 
   def index
-    authorize CaseContact
-
-    @current_organization_groups = current_organization_groups
-
-    @filterrific = initialize_filterrific(
-      all_case_contacts,
-      params[:filterrific],
-      select_options: {
-        sorted_by: CaseContact.options_for_sorted_by
-      }
-    ) || return
-
-    @pagy, @filtered_case_contacts = pagy(@filterrific.find)
-    case_contacts = CaseContact.case_hash_from_cases(@filtered_case_contacts)
-    case_contacts = case_contacts.select { |k, _v| current_user.casa_cases.pluck(:id).include?(k) } if current_user.volunteer?
-    case_contacts = case_contacts.select { |k, _v| k == params[:casa_case_id].to_i } if params[:casa_case_id].present?
-
-    @presenter = CaseContactPresenter.new(case_contacts)
+    load_case_contacts
   end
 
   def drafts
@@ -94,24 +79,6 @@ class CaseContactsController < ApplicationController
 
   def set_contact_types
     @contact_types = ContactType.for_organization(current_organization)
-  end
-
-  def current_organization_groups
-    current_organization.contact_type_groups
-      .includes(:contact_types)
-      .joins(:contact_types)
-      .where(contact_types: {active: true})
-      .uniq
-  end
-
-  def all_case_contacts
-    policy_scope(current_organization.case_contacts).preload(
-      :creator,
-      :followups,
-      contact_types: :contact_type_group,
-      contact_topic_answers: :contact_topic,
-      casa_case: :volunteers
-    )
   end
 
   def additional_expense_params
